Alteração atributos ID - Long

+++ src/main/java/com/marmitex/controller/TipoComidaController.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("tipo-comida")
-public class TipoComidaController extends CrudRestController<TipoComida, Integer> {
+public class TipoComidaController extends CrudRestController<TipoComida, Long> {
 
     private final TipoComidaService tipoComidaService;
 
@@ -22,7 +22,7 @@ public class TipoComidaController extends CrudRestController<TipoComida, Integer
     }
 
     @Override
-    public CrudService<TipoComida, Integer> getService() {
+    public CrudService<TipoComida, Long> getService() {
         return tipoComidaService;
     }
