+++ src/main/java/com/marmitex/data/ClienteData.java
@@ -3,6 +3,6 @@ package com.marmitex.data;
 import com.marmitex.model.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteData extends JpaRepository<Cliente, Integer> {
+public interface ClienteData extends JpaRepository<Cliente, Long> {
 
 }
