+++ src/main/java/com/marmitex/data/ComidaData.java
@@ -3,5 +3,5 @@ package com.marmitex.data;
 import com.marmitex.model.Comida;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComidaData extends JpaRepository<Comida, Integer> {
+public interface ComidaData extends JpaRepository<Comida, Long> {
 }
