+++ src/main/java/com/marmitex/service/ClienteService.java
@@ -3,5 +3,5 @@ package com.marmitex.service;
 import com.marmitex.framework.CrudService;
 import com.marmitex.model.Cliente;
 
-public interface ClienteService extends CrudService<Cliente, Integer> {
+public interface ClienteService extends CrudService<Cliente, Long> {
 }
