+++ src/main/java/com/marmitex/service/TipoComidaService.java
@@ -3,5 +3,5 @@ package com.marmitex.service;
 import com.marmitex.framework.CrudService;
 import com.marmitex.model.TipoComida;
 
-public interface TipoComidaService extends CrudService<TipoComida, Integer> {
+public interface TipoComidaService extends CrudService<TipoComida, Long> {
 }
